+++ drizzled/server_includes.h
@@ -576,7 +576,7 @@ int find_ref_key(KEY *key, uint key_count, uchar *record, Field *field,
                  uint *key_length, uint *keypart);
 void key_copy(uchar *to_key, uchar *from_record, KEY *key_info, uint key_length);
 void key_restore(uchar *to_record, uchar *from_key, KEY *key_info,
-                 uint key_length);
+                 uint16_t key_length);
 void key_zero_nulls(uchar *tuple, KEY *key_info);
 bool key_cmp_if_same(TABLE *form,const uchar *key,uint index,uint key_length);
 void key_unpack(String *to,TABLE *form,uint index);
